feat: refactor to re-use components
Changelog
==
 - refactors common code to be recycled between CLI and web
   server implementations
 - updates `README` to include how to run evaluator
